fix: section background not reaching page borders
Some sections (header, hero, quote and footer) should have a background
that covers the page horizontally. However, this was not being achieved
because of how the HTML was structured.

Now, each section is a div block (so it renders with normal flow) that
has `display: flex` and contains only one flex element: a div that has
the content for the section (and has a specific css class applied).
This way, the `section` parent is the one with the background, centers
the child, and sets the child size.